Added string_view class, using in Parser.
string_view is a lightweight 'view' of part or all of a string.
It doesn't own the string, just points to one, so the string/buffer it points
to needs to be around for >= the lifetime of the string view.

string_view will probably make it into the C++ standard library at some point
(right now it's in std::experimental namespace in some standard libraries)

Once it's standard we can remove this implementation.

Roughly based on
https://llvm.org/svn/llvm-project/libcxx/trunk/include/experimental/string_view

git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@7307 20aa11e4-6e6f-594c-a032-6291ace9ee65